Mimic regex return values for start and end of substring
Originally tried to use regex to return start and end indices but this
proved too annoying as re module ignores has a bunch of meta-characters
that I would need to escape. So the only change here is that the
return result is the same as if you used re.span (i.e. last index
number is not of the last character, but the one after that).