Refactor _is_abbreviation in hope of modest performance boost
It's unclear whether that will actually happen. Checking whether an
abbreviation exists using disjoint sets instead of linear search does
give a significant boost in timeit tests, but seeems to make no
difference when timing get_sentences on own or with nosetests. The big
time suck for long texts doesn't seem to be here, but elsewhere...