Refresh form should user navigate back from results page
More specifically, refresh the word/character counts and the Count
button should the user hit the Back button (or equivalent) from
Results page. Before the fix, word/character counts would be zero
even though there was a whole bunch of text in the textarea.

The fix is not great, a hack really. It relies on the pageshow event
which occurs whenever a page is shown after the first time its loaded.
In other words, _any_ navigation to the form page, not only hitting
the back button, will result in a refresh :-p.